fix: tsdx cjs bundle - move url.ts to cli folder so that dotenv is not included
