+++ package.json
@@ -4,7 +4,7 @@
   "description": "The front sdk for crema swap",
   "license": "MIT",
   "scripts": {
-    "build": "tsc",
+    "build": "tsdx build",
     "pre-swap": "ts-node ./cli/preSwap.ts",
     "create-pool": "ts-node ./cli/createPool.ts",
     "mint-pos": "ts-node ./cli/mintPosition.ts",
@@ -24,6 +24,7 @@
   "devDependencies": {
     "@types/chai": "^4.2.22",
     "chai": "^4.3.4",
+    "dotenv": "^16.0.0",
     "mocha": "^9.1.3",
     "ts-node": "^10.4.0",
     "tsdx": "^0.14.1",
@@ -34,5 +35,5 @@
   ],
   "typings": "dist/index.d.ts",
   "main": "dist/index.js",
-  "module": "dist/index.js"
+  "module": "dist/crema-sdk.esm.js"
 }
